% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backsolve.r
\docType{methods}
\name{backsolve}
\alias{backsolve}
\alias{backsolve,BaseLinAlg,spm-method}
\alias{backsolve,spm,BaseLinAlg-method}
\alias{backsolve,spm,spm-method}
\alias{forwardsolve,BaseLinAlg,spm-method}
\alias{forwardsolve,spm,BaseLinAlg-method}
\alias{forwardsolve,spm,spm-method}
\title{backsolve}
\usage{
\S4method{backsolve}{spm,spm}(r, x, k = ncol(r), upper.tri = TRUE,
  transpose = FALSE)

\S4method{backsolve}{spm,BaseLinAlg}(r, x, k = ncol(r), upper.tri = TRUE,
  transpose = FALSE)

\S4method{backsolve}{BaseLinAlg,spm}(r, x, k = ncol(r), upper.tri = TRUE,
  transpose = FALSE)

\S4method{forwardsolve}{spm,spm}(l, x, k = ncol(l), upper.tri = FALSE,
  transpose = FALSE)

\S4method{forwardsolve}{spm,BaseLinAlg}(l, x, k = ncol(l),
  upper.tri = FALSE, transpose = FALSE)

\S4method{forwardsolve}{BaseLinAlg,spm}(l, x, k = ncol(l),
  upper.tri = FALSE, transpose = FALSE)
}
\arguments{
\item{r, l}{A triangular coefficients matrix.}

\item{x}{The right hand sides.}

\item{k}{The number of equations (columns of r + rows of x) to use.}

\item{upper.tri}{Should the upper triangle be used? (if not the lower is)}

\item{transpose}{Should the transposed coefficients matrix be used? More efficient than
manually transposing with \code{t()}.}
}
\description{
Solve a triangular system.
}

