% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bind.r
\name{bind}
\alias{bind}
\alias{rbind.float32}
\alias{cbind.float32}
\title{rbind}
\usage{
\method{rbind}{float32}(..., deparse.level = 1)

\method{cbind}{float32}(..., deparse.level = 1)
}
\arguments{
\item{...}{vectors or matrices (numeric or float)}

\item{deparse.level}{ignored}
}
\value{
A matrix of the same type as the highest precision input.
}
\description{
\code{rbind()} and \code{cbind()} for floats.
}
\examples{
library(float)
x = fl(matrix(1:10, 5))

rbind(x, x)
cbind(x, x)

}
