% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/c.r
\docType{methods}
\name{c}
\alias{c}
\alias{c,spm-method}
\title{c}
\usage{
\S4method{c}{spm}(x, ...)
}
\arguments{
\item{x}{A float matrix.}

\item{...}{Additional elements (numeric/float vectors/matrices) to sum.}
}
\value{
TODO
}
\description{
Combine float/numeric vector(s)/matri[x|ces].
}
\examples{
\dontrun{
library(float)
x = flrunif(10, 3)

c(x, NA, 1L)
}

}

