% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/converters.r
\name{converters}
\alias{converters}
\alias{mat2spm}
\alias{fl}
\alias{spm2mat}
\alias{dbl}
\title{converters}
\usage{
mat2spm(x)

fl(x)

spm2mat(x)

dbl(x)
}
\arguments{
\item{x}{A numeric or float vector/matrix.}
}
\value{
The data stored in the type of whatever was asked for (the opposite of the
input).
}
\description{
Convert between a numeric vector/matrix and a float vector/matrix.
}
\details{
\code{fl()} is shorthand for \code{mat2spm()}; it behaves like a caster
(i.e., TO float). \code{dbl()} is shorthand for \code{spm2mat()}, casting
TO double (numeric matrix).
}
\examples{
\dontrun{
library(spm)

x = matrix(1:30, 10, 3)
s = fl(x) ### same as mat2spm(x)

y = dbl(s)

all.equal(x, y)
}

}
