% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extremes.r
\docType{methods}
\name{extremes}
\alias{extremes}
\alias{max,spm-method}
\alias{min,spm-method}
\title{extremes}
\usage{
\S4method{min}{spm}(x, ..., na.rm = FALSE)

\S4method{max}{spm}(x, ..., na.rm = FALSE)
}
\arguments{
\item{x}{A float matrix.}

\item{...}{Additional elements (numeric/float vectors/matrices) to sum.}

\item{na.rm}{should NA's be removed?}
}
\value{
A single value.
}
\description{
Min/max values for any combination of float/numeric vector(s)/matri[x|ces].
}
\details{
If there are any elements in \code{...}, all elements in the list will first
be summed in their native precision, then converted to double precision so
they can be combined with \code{base::sum()}.  The final result will be cast
to single precision if \code{...} contains only integer and/or float objects.
Otherwise, the return will be double precision.
}
\examples{
\dontrun{
library(spm)
x = flrunif(10, 3)

min(x)
min(x, 1)
}

}

