% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-constructor.r
\name{float}
\alias{float}
\title{float}
\usage{
float(length = 0, nrow, ncol)
}
\arguments{
\item{length}{Input data of type integer.}

\item{nrow, ncol}{Number of rows/columns if a matrix return is desired. See details section for
more information.}
}
\value{
A float vector/matrix of 0's.
}
\description{
An analogue to \code{integer()} and \code{double()} for preallocation.
}
\details{
If both of \code{nrow} and \code{ncol} are specified, then \code{length} is
ignored, and the retor is a matrix. If one (but not the other) of \code{nrow}
or \code{ncol} is given, then the function errors. Otherwise, a vector of
length \code{length} is returned.
}
\examples{
library(float)

float(10)
float(nrow=2, ncol=3)

}
