% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00-classes.r
\docType{class}
\name{float32-class}
\alias{float32-class}
\title{Class float32}
\description{
An S4 container for single precision (float) vector/matrix objects.
}
\section{Slots}{

\describe{
\item{\code{ptr}}{An external pointer to a struct.}
}}

