% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-constructor.r
\name{float32}
\alias{float32}
\title{float32}
\usage{
float32(x)
}
\arguments{
\item{x}{Input data of type integer.}
}
\value{
A \code{float32} class object.
}
\description{
A \code{float32} class constructor. For developers only.
}
\details{
Wraps the integer-type data in the \code{float32} S4 class, so that the data
will be interpreted as 32-bit floats.

If instead you merely want to convert numeric/double data to float type,
instead you should call \code{fl(x)}.
}

