% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/math.r
\docType{methods}
\name{hyperbolic}
\alias{acosh,spm-method}
\alias{asinh,spm-method}
\alias{atanh,spm-method}
\alias{cosh,spm-method}
\alias{hyperbolic}
\alias{sinh,spm-method}
\alias{tanh,spm-method}
\title{Hyperbolic functions}
\usage{
\S4method{sinh}{spm}(x)

\S4method{cosh}{spm}(x)

\S4method{tanh}{spm}(x)

\S4method{asinh}{spm}(x)

\S4method{acosh}{spm}(x)

\S4method{atanh}{spm}(x)
}
\arguments{
\item{x}{A float vector/matrix.}
}
\value{
A float vector/matrix of the same dimensions as the input.
}
\description{
Hyperbolic functions.
}
\examples{
\dontrun{
library(float)

x = flrunif(10)
sinh(x)
}

}

