% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is.float.r
\name{is.float}
\alias{is.float}
\title{is.float}
\usage{
is.float(x)
}
\arguments{
\item{x}{An R object.}
}
\value{
A logical value.
}
\description{
Tests if argument is a float matrix.
}
\details{
\code{is.float()} and \code{is.float()} are different names for the same
function.
}
\examples{
library(float)

x = matrix(0, 5, 5)
s = flrunif(10, 3)
is.float(x)
is.float(s)

}
