% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isSymmetric.r
\name{isSymmetric}
\alias{isSymmetric}
\title{isSymmetric}
\arguments{
\item{object}{A float vector/matrix.}
}
\value{
A logical value.
}
\description{
Test if a float matrix is symmetric.
}
\examples{
\dontrun{
library(spm)

s = flrunif(10, 3)
isSymmetric(s)

cp = crossprod(s)
isSymmetric(s)
}

}
