% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log.r
\docType{methods}
\name{log}
\alias{log}
\alias{exp,spm-method}
\alias{expm1,spm-method}
\alias{log,spm-method}
\alias{log10,spm-method}
\alias{log2,spm-method}
\title{Logarithms and Exponentials}
\usage{
\S4method{exp}{spm}(x)

\S4method{expm1}{spm}(x)

\S4method{log}{spm}(x, base = exp(1))

\S4method{log10}{spm}(x)

\S4method{log2}{spm}(x)
}
\arguments{
\item{x}{A float vector/matrix.}

\item{base}{The logarithm base.}
}
\value{
A float vector/matrix of the same dimensions as the input.
}
\description{
exp/log functions.
}
\examples{
\dontrun{
library(spm)

x = flrunif(10)
log(x)
}

}
