% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/math.r
\docType{methods}
\name{miscmath}
\alias{abs,spm-method}
\alias{miscmath}
\alias{sqrt,spm-method}
\title{Miscellaneous mathematical functions}
\usage{
\S4method{abs}{spm}(x)

\S4method{sqrt}{spm}(x)
}
\arguments{
\item{x}{A float vector/matrix.}
}
\value{
A float vector/matrix of the same dimensions as the input.
}
\description{
Miscellaneous mathematical functions.
}
\examples{
\dontrun{
library(float)

x = flrunif(10)
sqrt(x)
}

}

