% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/na.r
\docType{methods}
\name{na}
\alias{na}
\alias{is.na,spm-method}
\alias{na.omit,spm-method}
\alias{na.exclude,spm-method}
\title{NA}
\usage{
\S4method{is.na}{spm}(x)

\S4method{na.omit}{spm}(object, ...)

\S4method{na.exclude}{spm}(object, ...)
}
\arguments{
\item{x, object}{A float vector/matrix.}

\item{...}{Ignored.}
}
\description{
NA utilities.
}
\examples{
\dontrun{
library(spm)

s = flrunif(10, 3)
is.na(s)
}

}
