% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/names.r
\docType{methods}
\name{names}
\alias{names}
\alias{names,float32-method}
\alias{names<-,float32-method}
\alias{rownames,float32-method}
\alias{rownames<-,float32-method}
\alias{colnames,float32-method}
\alias{colnames<-,float32-method}
\alias{dimnames,float32-method}
\alias{dimnames<-,float32-method}
\title{names}
\usage{
\S4method{names}{float32}(x)

\S4method{names}{float32}(x) <- value

\S4method{rownames}{float32}(x)

\S4method{rownames}{float32}(x) <- value

\S4method{colnames}{float32}(x)

\S4method{colnames}{float32}(x) <- value

\S4method{dimnames}{float32}(x)

\S4method{dimnames}{float32}(x) <- value
}
\arguments{
\item{x}{A float vector/matrix.}

\item{value}{Replacement value.}
}
\description{
"name" setter/getters.
}
