% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/norm.r
\docType{methods}
\name{norm}
\alias{norm}
\alias{norm,spm,ANY-method}
\title{norm}
\usage{
\S4method{norm}{spm,ANY}(x, type = c("O", "I", "F", "M", "2"))
}
\arguments{
\item{x}{A float vector/matrix.}

\item{type}{"O"-ne, "I"-nfinity, "F"-robenius, "M"-ax modulus, and "2" norms.}
}
\value{
A single float.
}
\description{
Compute matrix norm.
}
\examples{
\dontrun{
library(spm)

s = flrunif(10, 3)
norm(s)
}

}

