% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.r
\docType{methods}
\name{print-float32}
\alias{print,float32-method}
\alias{print-float32}
\alias{show,float32-method}
\title{print-float32}
\usage{
\S4method{print}{float32}(x, ...)

\S4method{show}{float32}(object)
}
\arguments{
\item{x, object}{A float vector/matrix.}

\item{...}{Ignored.}
}
\description{
Print methods for float vector/matrices.
}
\examples{
library(float)

s = flrunif(10, 3)
print(s)
s

}

