% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qr.r
\docType{methods}
\name{qr}
\alias{qr}
\alias{qr,spm-method}
\alias{qr.Q,ANY-method}
\alias{qr.R,ANY-method}
\title{QR}
\usage{
\S4method{qr}{spm}(x, tol = 1e-07, ...)

\S4method{qr.Q}{ANY}(qr, complete = FALSE, Dvec)

\S4method{qr.R}{ANY}(qr, complete = FALSE)
}
\arguments{
\item{x}{A float matrix.}

\item{tol}{The tolerance for determining numerical column rank.}

\item{...}{Ignored.}

\item{qr}{Output of \code{qr()}.}

\item{complete}{TODO}

\item{Dvec}{Ignored}
}
\description{
QR factorization and related functions.
}
\details{
The factorization is performed by the LAPACK routine \code{sgeqp3()}.
}
\examples{
\dontrun{
library(spm)

x = flrunif(10, 3)
qr(x)
}

}

