% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rep.r
\name{rep}
\alias{rep}
\alias{rep.float32}
\title{rep}
\usage{
\method{rep}{float32}(x, ...)
}
\arguments{
\item{x}{A float matrix.}

\item{...}{Additional elements (numeric/float vectors/matrices) to sum.}
}
\value{
A matrix of the same type as the highest precision input.
}
\description{
Combine float/numeric vector(s)/matri[x|ces].
}
\examples{
library(float)
x = flrunif(10, 3)

c(x, NA, 1L)

}

