% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/math.r
\docType{methods}
\name{round}
\alias{ceiling,spm-method}
\alias{floor,spm-method}
\alias{round}
\alias{round,spm-method}
\alias{trunc,spm-method}
\title{Round}
\usage{
\S4method{ceiling}{spm}(x)

\S4method{floor}{spm}(x)

\S4method{trunc}{spm}(x, ...)

\S4method{round}{spm}(x, digits = 0)
}
\arguments{
\item{x}{A float vector/matrix.}

\item{...}{ignored}

\item{digits}{The number of digits to use in rounding.}
}
\value{
A float vector/matrix of the same dimensions as the input.
}
\description{
Rounding functions.
}
\examples{
\dontrun{
library(float)

x = flrnorm(10)
floor(x)
}

}

