% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale.r
\name{scale}
\alias{scale}
\alias{scale,float32-method}
\title{scale}
\usage{
\S4method{scale}{float32}(x, center = TRUE, scale = TRUE)
}
\arguments{
\item{x}{A float vector/matrix.}

\item{center, scale}{Logical}
}
\value{
A float matrix.
}
\description{
Center/scale a float vector/matrix.
}
\details{
Only logical \code{center} and \code{scale} parameters are accepted at this
time.
}
\examples{
library(float)

s = flrunif(10, 3)
scale(s)

}
