% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sign.r
\docType{methods}
\name{sign}
\alias{sign}
\alias{sign,float32-method}
\title{sign}
\usage{
\S4method{sign}{float32}(x)
}
\arguments{
\item{x}{A float vector/matrix.}
}
\value{
A float vector/matrix.
}
\description{
Sign.
}
