% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sum.r
\docType{methods}
\name{sum}
\alias{sum}
\alias{sum,spm-method}
\title{sum}
\usage{
\S4method{sum}{spm}(x, ..., na.rm = FALSE)
}
\arguments{
\item{x}{}

\item{...}{Additional elements (shaqs or numeric vectors/matrices) to sum.  Regular
vectors/matrices are assumed to be global, but will only be included to the
sum on rank 0.}

\item{na.rm}{should NA's be removed?}
}
\value{
A single value.
}
\description{
Sums the elements of a shaq.
}
\details{
TODO
}
\examples{
\dontrun{
library(spm)
x = flrunif(10, 3)

sum(x)
sum(x, 1)
}

}

