% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svd.r
\docType{methods}
\name{svd}
\alias{La.svd,spm-method}
\alias{svd}
\alias{svd,spm-method}
\title{SVD}
\usage{
\S4method{La.svd}{spm}(x, nu = min(n, p), nv = min(n, p))

\S4method{svd}{spm}(x, nu = min(n, p), nv = min(n, p), LINPACK = FALSE)
}
\arguments{
\item{x}{A float matrix.}

\item{nu, nv}{The number of left/right singular vectors to return.}

\item{LINPACK}{Ignored}
}
\description{
SVD factorization.
}
\details{
The factorization is performed by the LAPACK routine \code{sgesdd()}.
}
\examples{
\dontrun{
library(float)

x = flrunif(10, 3)
svd(x)
}

}

