% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale.r, R/sweep.r
\docType{methods}
\name{sweep}
\alias{scale,spm-method}
\alias{sweep}
\alias{sweep,spm-method}
\title{sweep}
\usage{
\S4method{scale}{spm}(x, center = TRUE, scale = TRUE)

\S4method{sweep}{spm}(x, MARGIN, STATS, FUN = "-", check.margin = TRUE, ...)
}
\arguments{
\item{x}{A float vector/matrix.}

\item{MARGIN}{1 (rows) or 2 (columns)}

\item{STATS}{Vector to sweep out.}

\item{FUN}{Sweeping function; must be one of \code{"+"}, \code{"-"}, \code{"*"},  or
\code{"/"}.}

\item{check.margin}{Should x/STATS margin lengths be checked?}

\item{...}{Theoretically these are additional arguments passed to an arbitrary function.
However, we only support basic arithmetic, so they are ignored.}

\item{x}{A float vector/matrix.}

\item{MARGIN}{1 (rows) or 2 (columns)}

\item{STATS}{Vector to sweep out.}

\item{FUN}{Sweeping function; must be one of \code{"+"}, \code{"-"}, \code{"*"},  or
\code{"/"}.}

\item{check.margin}{Should x/STATS margin lengths be checked?}

\item{...}{Theoretically these are additional arguments passed to an arbitrary function.
However, we only support basic arithmetic, so they are ignored.}
}
\value{
A matrix of the same type as the highest precision input.

A matrix of the same type as the highest precision input.
}
\description{
Extract subsets of a float vector/matrix.

Sweep a vector through a float matrix.
}
\details{
Note that if the length of \code{STATS} does not recycle exactly across
\code{MARGIN}, the results here will differ slightly from the results of
base R.

Note that if the length of \code{STATS} does not recycle exactly across
\code{MARGIN}, the results here will differ slightly from the results of
base R.
}
\examples{
\dontrun{
library(spm)

s = flrunif(10, 3)
sweep(s, 2, fl(1))
}

\dontrun{
library(spm)

s = flrunif(10, 3)
sweep(s, 2, fl(1))
}

}

