% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/math.r
\docType{methods}
\name{trig}
\alias{trig}
\alias{sin,spm-method}
\alias{cos,spm-method}
\alias{tan,spm-method}
\alias{asin,spm-method}
\alias{acos,spm-method}
\alias{atan,spm-method}
\alias{sinh,spm-method}
\alias{cosh,spm-method}
\alias{tanh,spm-method}
\alias{asinh,spm-method}
\alias{acosh,spm-method}
\alias{atanh,spm-method}
\title{Trigonometric functions}
\usage{
\S4method{sin}{spm}(x)

\S4method{cos}{spm}(x)

\S4method{tan}{spm}(x)

\S4method{asin}{spm}(x)

\S4method{acos}{spm}(x)

\S4method{atan}{spm}(x)

\S4method{sinh}{spm}(x)

\S4method{cosh}{spm}(x)

\S4method{tanh}{spm}(x)

\S4method{asinh}{spm}(x)

\S4method{acosh}{spm}(x)

\S4method{atanh}{spm}(x)
}
\arguments{
\item{x}{A float vector/matrix.}
}
\value{
A float vector/matrix of the same dimensions as the input.
}
\description{
Basic trig functions.
}
\examples{
\dontrun{
library(spm)

x = flrunif(10)
sin(x)
}

}
