% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xpose.r
\docType{methods}
\name{xpose}
\alias{t,spm-method}
\alias{xpose}
\title{xpose}
\usage{
\S4method{t}{spm}(x)
}
\arguments{
\item{x}{A float vector/matrix.}
}
\value{
A float vector/matrix.
}
\description{
Transpose a float vector/matrix.
}
\examples{
\dontrun{
library(float)

s = flrunif(10, 3)
dim(s)
ts = t(s)
dim(ts)
}

}

