% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arithmetic.r
\docType{methods}
\name{arithmetic}
\alias{*,LinAlg,LinAlg-method}
\alias{+,LinAlg,LinAlg-method}
\alias{-,LinAlg,LinAlg-method}
\alias{/,LinAlg,LinAlg-method}
\alias{^,LinAlg,LinAlg-method}
\alias{arithmetic}
\title{arithmetic}
\usage{
\S4method{+}{LinAlg,LinAlg}(e1, e2)

\S4method{*}{LinAlg,LinAlg}(e1, e2)

\S4method{-}{LinAlg,LinAlg}(e1, e2)

\S4method{/}{LinAlg,LinAlg}(e1, e2)

\S4method{^}{LinAlg,LinAlg}(e1, e2)
}
\arguments{
\item{e1, e2}{Numeric/float vectors/matrices.}
}
\value{
A matrix of the same type as the highest precision input.
}
\description{
Binary arithmetic numeric/float matrices.
}
\examples{
\dontrun{
library(spm)

s1 = flrunif(5, 5)
s2 = flrunif(5, 2)
x = matrix(1:25, 5)

s1 + s2 # float

storage.mode(x) # integer
x + s2 # float

storage.mode(x) = "double"
x + s2 # double
}

}

