% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binary.r
\docType{methods}
\name{arithmetic}
\alias{*,BaseLinAlg,spm-method}
\alias{*,spm,BaseLinAlg-method}
\alias{*,spm,spm-method}
\alias{+,BaseLinAlg,spm-method}
\alias{+,spm,BaseLinAlg-method}
\alias{+,spm,spm-method}
\alias{-,BaseLinAlg,spm-method}
\alias{-,spm,BaseLinAlg-method}
\alias{-,spm,spm-method}
\alias{/,BaseLinAlg,spm-method}
\alias{/,spm,BaseLinAlg-method}
\alias{/,spm,spm-method}
\alias{<,BaseLinAlg,spm-method}
\alias{<,spm,BaseLinAlg-method}
\alias{<,spm,spm-method}
\alias{<=,BaseLinAlg,spm-method}
\alias{<=,spm,BaseLinAlg-method}
\alias{<=,spm,spm-method}
\alias{==,BaseLinAlg,spm-method}
\alias{==,spm,BaseLinAlg-method}
\alias{==,spm,spm-method}
\alias{>,BaseLinAlg,spm-method}
\alias{>,spm,BaseLinAlg-method}
\alias{>,spm,spm-method}
\alias{>=,BaseLinAlg,spm-method}
\alias{>=,spm,BaseLinAlg-method}
\alias{>=,spm,spm-method}
\alias{^,BaseLinAlg,spm-method}
\alias{^,spm,BaseLinAlg-method}
\alias{^,spm,spm-method}
\alias{arithmetic}
\title{arithmetic}
\usage{
\S4method{+}{spm,spm}(e1, e2)

\S4method{*}{spm,spm}(e1, e2)

\S4method{-}{spm,spm}(e1, e2)

\S4method{/}{spm,spm}(e1, e2)

\S4method{^}{spm,spm}(e1, e2)

\S4method{+}{spm,BaseLinAlg}(e1, e2)

\S4method{*}{spm,BaseLinAlg}(e1, e2)

\S4method{-}{spm,BaseLinAlg}(e1, e2)

\S4method{/}{spm,BaseLinAlg}(e1, e2)

\S4method{^}{spm,BaseLinAlg}(e1, e2)

\S4method{+}{BaseLinAlg,spm}(e1, e2)

\S4method{*}{BaseLinAlg,spm}(e1, e2)

\S4method{-}{BaseLinAlg,spm}(e1, e2)

\S4method{/}{BaseLinAlg,spm}(e1, e2)

\S4method{^}{BaseLinAlg,spm}(e1, e2)

\S4method{<}{spm,spm}(e1, e2)

\S4method{<=}{spm,spm}(e1, e2)

\S4method{==}{spm,spm}(e1, e2)

\S4method{>}{spm,spm}(e1, e2)

\S4method{>=}{spm,spm}(e1, e2)

\S4method{<}{spm,BaseLinAlg}(e1, e2)

\S4method{<=}{spm,BaseLinAlg}(e1, e2)

\S4method{==}{spm,BaseLinAlg}(e1, e2)

\S4method{>}{spm,BaseLinAlg}(e1, e2)

\S4method{>=}{spm,BaseLinAlg}(e1, e2)

\S4method{<}{BaseLinAlg,spm}(e1, e2)

\S4method{<=}{BaseLinAlg,spm}(e1, e2)

\S4method{==}{BaseLinAlg,spm}(e1, e2)

\S4method{>}{BaseLinAlg,spm}(e1, e2)

\S4method{>=}{BaseLinAlg,spm}(e1, e2)
}
\arguments{
\item{e1, e2}{Numeric/float vectors/matrices.}
}
\value{
A matrix of the same type as the highest precision input.
}
\description{
Binary arithmetic numeric/float matrices.
}
\examples{
\dontrun{
library(float)

s1 = flrunif(5, 5)
s2 = flrunif(5, 2)
x = matrix(1:25, 5)

s1 + s2 # float

storage.mode(x) # integer
x + s2 # float

storage.mode(x) = "double"
x + s2 # double
}

}

