% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bracket.r
\docType{methods}
\name{bracket}
\alias{bracket}
\alias{[,spm-method}
\title{Extract}
\usage{
\S4method{[}{spm}(x, i, j, drop = TRUE)
}
\arguments{
\item{x}{A float vector/matrix.}

\item{i, j}{The indices. NOTE that at this time only non-specified and positive indexed
i/j are accepted.}

\item{drop}{Logical. If TRUE, single column matrices will be treated as one-dimensional
vectors.}
}
\value{
A float vector/matrix.
}
\description{
Extract subsets of a float vector/matrix.
}
\examples{
\dontrun{
library(spm)

s = flrunif(10, 3)
s[, -1]
s[c(1, 3, 5, 7), 1:2]
}

}
