% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chol.r
\docType{methods}
\name{chol}
\alias{chol}
\alias{chol,spm-method}
\title{chol}
\usage{
\S4method{chol}{spm}(x)
}
\arguments{
\item{x}{A float vector/matrix.}
}
\value{
A float vector/matrix.
}
\description{
Extract subsets of a float vector/matrix.
}
\examples{
\dontrun{
library(spm)

s = flrunif(10, 3)
cp = crossprod(s)
chol(cp)
}

}

