% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chol.r
\docType{methods}
\name{chol}
\alias{chol}
\alias{chol,float32-method}
\title{chol}
\usage{
\S4method{chol}{float32}(x)
}
\arguments{
\item{x}{A float vector/matrix.}
}
\value{
A float vector/matrix.
}
\description{
Cholesky factorization for a float vector/matrix.
}
\examples{
library(float)

s = flrunif(10, 3)
cp = crossprod(s)
chol(cp)

}
