% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/converters.r
\name{converters}
\alias{converters}
\alias{dbl}
\alias{fl}
\title{converters}
\usage{
fl(x, strict = FALSE)

dbl(x, strict = FALSE)
}
\arguments{
\item{x}{A numeric or float vector/matrix.}

\item{strict}{Should the function error if given the wrong kind of input? Otherwise it just
silently returns the input.}
}
\value{
The data stored in the type of whatever was asked for (the opposite of the
input).
}
\description{
Convert between a numeric vector/matrix and a float vector/matrix.
}
\examples{
\dontrun{
library(float)

x = matrix(1:30, 10, 3)
s = fl(x)

y = dbl(s)

all.equal(x, y)
}

}

