% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dims.r
\docType{methods}
\name{dims}
\alias{dim,spm-method}
\alias{dims}
\alias{length,spm-method}
\alias{ncol,spm-method}
\alias{nrow,spm-method}
\title{dim}
\usage{
\S4method{nrow}{spm}(x)

\S4method{ncol}{spm}(x)

\S4method{dim}{spm}(x)

\S4method{length}{spm}(x)
}
\arguments{
\item{x}{A float vector/matrix.}
}
\value{
The requested integer values.
}
\description{
Dimension information for a float vector/matrix.
}
\examples{
\dontrun{
library(spm)

s = flrunif(10, 3)
dim(s)
nrow(s)
ncol(s)
}

}

