% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dims.r
\name{dims}
\alias{dims}
\alias{nrow,float32-method}
\alias{ncol,float32-method}
\alias{NROW,float32-method}
\alias{NCOL,float32-method}
\alias{dim,float32-method}
\alias{length,float32-method}
\alias{dim<-,float32-method}
\title{dim}
\usage{
\S4method{nrow}{float32}(x)

\S4method{ncol}{float32}(x)

\S4method{NROW}{float32}(x)

\S4method{NCOL}{float32}(x)

\S4method{dim}{float32}(x)

\S4method{length}{float32}(x)

\S4method{dim}{float32}(x) <- value
}
\arguments{
\item{x}{A float vector/matrix.}

\item{value}{The right hand side for the "setter" (\code{dim<-}).}
}
\value{
The requested integer values.
}
\description{
Dimension information for a float vector/matrix.
}
\examples{
library(float)

s = flrunif(10, 3)
dim(s)
nrow(s)
ncol(s)

}
