% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/math.r
\name{hyperbolic}
\alias{hyperbolic}
\title{Hyperbolic functions}
\arguments{
\item{x}{A float vector/matrix.}
}
\value{
A float vector/matrix of the same dimensions as the input.
}
\description{
Hyperbolic functions.
}
\examples{
\dontrun{
library(spm)

x = flrunif(10)
sinh(x)
}

}
