% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/math.r
\docType{methods}
\name{hyperbolic}
\alias{hyperbolic}
\alias{sinh,float32-method}
\alias{cosh,float32-method}
\alias{tanh,float32-method}
\alias{asinh,float32-method}
\alias{acosh,float32-method}
\alias{atanh,float32-method}
\title{Hyperbolic functions}
\usage{
\S4method{sinh}{float32}(x)

\S4method{cosh}{float32}(x)

\S4method{tanh}{float32}(x)

\S4method{asinh}{float32}(x)

\S4method{acosh}{float32}(x)

\S4method{atanh}{float32}(x)
}
\arguments{
\item{x}{A float vector/matrix.}
}
\value{
A float vector/matrix of the same dimensions as the input.
}
\description{
Hyperbolic functions.
}
\examples{
\dontrun{
library(float)

x = flrunif(10)
sinh(x)
}

}
