% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is.spm.r
\name{is.spm}
\alias{is.spm}
\title{is.spm}
\usage{
is.spm(x)
}
\arguments{
\item{x}{An R object.}
}
\value{
A logical value.
}
\description{
Tests if argument is a float matrix.
}
\examples{
\dontrun{
library(spm)

x = matrix(0, 5, 5)
s = flrunif(10, 3)
is.spm(x)
is.spm(s)
}

}
