% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isSymmetric.r
\docType{methods}
\name{isSymmetric}
\alias{isSymmetric}
\alias{isSymmetric,float32-method}
\title{isSymmetric}
\usage{
\S4method{isSymmetric}{float32}(object, ...)
}
\arguments{
\item{object}{A float vector/matrix.}

\item{...}{Ignored.}
}
\value{
A logical value.
}
\description{
Test if a float matrix is symmetric.
}
\examples{
library(float)

s = flrunif(10, 3)
isSymmetric(s)

cp = crossprod(s)
isSymmetric(s)

}

