% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/math.r
\docType{methods}
\name{mathis}
\alias{mathis}
\alias{is.finite,float32-method}
\alias{is.infinite,float32-method}
\alias{is.nan,float32-method}
\title{Finite, infinite, and NaNs}
\usage{
\S4method{is.finite}{float32}(x)

\S4method{is.infinite}{float32}(x)

\S4method{is.nan}{float32}(x)
}
\arguments{
\item{x}{A float vector/matrix.}
}
\value{
An integer vector/matrix of the same dimensions as the input.
}
\description{
Finite, infinite, and NaNs.
}
\examples{
\dontrun{
library(float)

x = flrnorm(10)
is.nan(sqrt(x))
}

}
