% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/na.r
\docType{methods}
\name{na}
\alias{is.na,float32-method}
\alias{na}
\alias{na.exclude,float32-method}
\alias{na.omit,float32-method}
\title{NA}
\usage{
\S4method{is.na}{float32}(x)

\S4method{na.omit}{float32}(object, ...)

\S4method{na.exclude}{float32}(object, ...)
}
\arguments{
\item{x, object}{A float vector/matrix.}

\item{...}{Ignored.}
}
\description{
NA utilities.
}
\examples{
library(float)

s = flrunif(10, 3)
is.na(s)

}

