% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numbytes.r
\docType{methods}
\name{object.size}
\alias{object.size}
\alias{object.size,spm-method}
\title{object.size}
\usage{
\S4method{object.size}{spm}(x)
}
\arguments{
\item{x}{A float vector/matrix.}
}
\description{
Return the memory being used to store the object.
}
\examples{
\dontrun{
library(spm)

s = flrunif(10, 3)
object.size(s)

### if the memuse package is also available:
library(memuse)
memuse(s)
}

}
