% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.r
\docType{methods}
\name{print-spm}
\alias{print-spm}
\alias{print,spm-method}
\alias{show,spm-method}
\title{print-spm}
\usage{
\S4method{print}{spm}(x, ...)

\S4method{show}{spm}(object)
}
\arguments{
\item{x}{A float vector/matrix.}
}
\description{
Print methods for float vector/matrices.
}
\examples{
\dontrun{
library(spm)

s = flrunif(10, 3)
print(s)
s
}

}
