% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rand.r
\name{rand}
\alias{flrnorm}
\alias{flrunif}
\alias{rand}
\title{Generators}
\usage{
flrunif(m, n, min = 0, max = 1)

flrnorm(m, n, mean = 0, sd = 1)
}
\arguments{
\item{m, n}{The dimensions of the matrix/vector. \code{m} must be specified.  If \code{n}
is not, then the return is a vector.}

\item{min, max}{Minimum and maximum values for the uniform generator.}

\item{mean, sd}{Mean and standard deviation values for the normal generator.}
}
\description{
Random float vector/matrix generators. \code{flrunif()} produces uniform
random values. \code{flrnorm()} produces random normal values.
}
\details{
The data is produced without a double precision copy.  That is, it is not
(computationally) equivalent to \code{fl(matrix(runif(...)))}, though the
operations are conceptually the same.

To produce a vector instead of a matrix, leave argument \code{n} blank.
Setting \code{n=1} will produce an mx1 matrix.
}
\examples{
\dontrun{
library(float)

flrunif(10) # length 10 vector
flrunif(10, 1) # 10x1 matrix
flrunif(10, min=10, max=20)
}

}

