% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cond.r
\docType{methods}
\name{rcond}
\alias{rcond}
\alias{rcond,float32-method}
\title{rcond}
\usage{
\S4method{rcond}{float32}(x, norm = c("O", "I", "1"), triangular = FALSE,
  ...)
}
\arguments{
\item{x}{A float vector/matrix.}

\item{norm}{"O"-ne or "I"-nfinity norm.}

\item{triangular}{Should only the lower triangle be used?}

\item{...}{Additional arguments.}
}
\value{
A single float.
}
\description{
Compute matrix norm.
}
\examples{
library(float)

s = flrunif(10, 3)
rcond(s)

}

