% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rep.r
\name{rep}
\alias{rep}
\alias{rep.float32}
\title{rep}
\usage{
\method{rep}{float32}(x, ...)
}
\arguments{
\item{x}{A float matrix.}

\item{...}{Additional arguments (passed to \code{base::rep}).}
}
\value{
A float vector.
}
\description{
Replicate elements of a float vector/matrix.
}
\examples{
library(float)
x = fl(matrix(1:6, 3, 2))

rep(x, 5)

}
