% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/math.r
\name{round}
\alias{round}
\alias{ceiling,float32-method}
\alias{floor,float32-method}
\alias{trunc,float32-method}
\alias{round,float32-method}
\title{Round}
\usage{
\S4method{ceiling}{float32}(x)

\S4method{floor}{float32}(x)

\S4method{trunc}{float32}(x, ...)

\S4method{round}{float32}(x, digits = 0)
}
\arguments{
\item{x}{A float vector/matrix.}

\item{...}{ignored}

\item{digits}{The number of digits to use in rounding.}
}
\value{
A float vector/matrix of the same dimensions as the input.
}
\description{
Rounding functions.
}
\examples{
library(float)

x = flrnorm(10)
floor(x)

}
