% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/math.r
\docType{methods}
\name{specialmath}
\alias{specialmath}
\alias{gamma,float32-method}
\alias{lgamma,float32-method}
\title{Special mathematical functions}
\usage{
\S4method{gamma}{float32}(x)

\S4method{lgamma}{float32}(x)
}
\arguments{
\item{x}{A float vector/matrix.}
}
\value{
A float vector/matrix of the same dimensions as the input.
}
\description{
Special mathematical functions.
}
\examples{
\dontrun{
library(float)

x = flrunif(10)
lgamma(x)
}

}
