% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/math.r
\docType{methods}
\name{trig}
\alias{acos,spm-method}
\alias{asin,spm-method}
\alias{atan,spm-method}
\alias{cos,spm-method}
\alias{sin,spm-method}
\alias{tan,spm-method}
\alias{trig}
\title{Trigonometric functions}
\usage{
\S4method{sin}{spm}(x)

\S4method{cos}{spm}(x)

\S4method{tan}{spm}(x)

\S4method{asin}{spm}(x)

\S4method{acos}{spm}(x)

\S4method{atan}{spm}(x)
}
\arguments{
\item{x}{A float vector/matrix.}
}
\value{
A float vector/matrix of the same dimensions as the input.
}
\description{
Basic trig functions.
}
\examples{
\dontrun{
library(float)

x = flrunif(10)
sin(x)
}

}

