% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/math.r
\name{trig}
\alias{trig}
\alias{sin,float32-method}
\alias{cos,float32-method}
\alias{tan,float32-method}
\alias{asin,float32-method}
\alias{acos,float32-method}
\alias{atan,float32-method}
\title{Trigonometric functions}
\usage{
\S4method{sin}{float32}(x)

\S4method{cos}{float32}(x)

\S4method{tan}{float32}(x)

\S4method{asin}{float32}(x)

\S4method{acos}{float32}(x)

\S4method{atan}{float32}(x)
}
\arguments{
\item{x}{A float vector/matrix.}
}
\value{
A float vector/matrix of the same dimensions as the input.
}
\description{
Basic trig functions.
}
\examples{
\dontrun{
library(float)

x = flrunif(10)
sin(x)
}

}
