% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xpose.r
\name{xpose}
\alias{xpose}
\alias{t,float32-method}
\title{xpose}
\usage{
\S4method{t}{float32}(x)
}
\arguments{
\item{x}{A float vector/matrix.}
}
\value{
A float vector/matrix.
}
\description{
Transpose a float vector/matrix.
}
\examples{
library(float)

s = flrunif(10, 3)
dim(s)
ts = t(s)
dim(ts)

}
